Commit 34.2 - Gauna Nicolas (30/08/2024)

+++ Project_Triathlon/src/model/athlete/Athlete.java
@@ -263,7 +263,7 @@ public abstract class Athlete {
 	}
 	//------------------------------------------------>||CLASS METHODS||<--------------------------------------------------------\\
 
-	public double getSpeed() {
+	public int getSpeed() {
 		return 500 - (userSpeedAdjustment-5) * 50;
 	}
 	public int getPositionChange(Race race){
